refactor(log): move log message formatting into shared.ts
This way it can be used by other logging namespaces aswell
