+++ src/lib/log/index.ts
@@ -1,7 +1,9 @@
+import * as driver from "./Driver";
 import * as serial from "./Serial";
 
 const logger = Object.freeze({
 	serial,
+	driver,
 });
 
 export default logger;
